Delete runonly api
